chore: move submit customer logic to separate file
